backend copy/requirements.txt
pandas
gspread
gspread-dataframe
mysql-connector-python
python-dotenv
